jank abstracted sim_traj version
